fix(iam): enable facebook verification with non-owner/developer accounts
This fixes a bug where Facebook stamp verification fails when used with a Facebook account that is
not associated with the Gitcoin Passport Facebook app development team. The root cause is that an
app token (or similarly privileged user token from a dev team user) is necessary to access the
Facebook Graph API we are using to inspect details about the /user/ access token provided by the
Facebook login on the front end. Therefore, under normal usage, the app access token and the user
access token are two different token values; the bug is that we were attempting to use the user
access token for both purposes. This was not caught during development/testing/initial acceptance
because the developer Facebook account was used, not a 'normal' Facebook account.

[fixes #116]