test(database-client): generate a new random DID for ceramic integration tests
Ceramic integration tests are now consistently timing out (and failing). It seems to be due to the
enormous number of commits logged when using the same DID for all test runs that causes tests to
time out. Replacing this with a newly generated DID each time results in fast-enough integration
tests that pass more consistently.